requirements.txt
Django==4.1.4
djangorestframework==3.14.0
python-decouple==3.6
psycopg2==2.9.5
Pillow==9.3.0
django-cleanup==6.0.0
